% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_airway.R
\docType{data}
\name{se_airway}
\alias{se_airway}
\title{Example SummarizedExperiment Dataset}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{A matrix of scaled counts}
\item{colData}{Sample metadata (cell type, treatment, etc.)}
}
}
\source{
Airway dataset
}
\usage{
se_airway
}
\description{
This dataset contains scaled RNA-seq counts for airway samples.
}
\keyword{datasets}
