% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_airway.R
\docType{data}
\name{se_airway}
\alias{se_airway}
\title{Example SummarizedExperiment dataset}
\format{
A SummarizedExperiment object with:
\describe{
\item{assays}{A matrix of scaled counts}
\item{colData}{Sample metadata (cell type, treatment, etc.)}
}
}
\source{
airway package
}
\usage{
data(se_airway)
}
\description{
Scaled RNA-seq counts for airway samples packaged as a
\link[SummarizedExperiment]{SummarizedExperiment}.
}
\keyword{datasets}
